added port to prometheus config
